requirements.txt
flask==2.2.5
openai==0.28.1
python-dotenv==0.21.1
